refactor(components): flatten directories
