--- src/components/common/OccurrenceChip/index.ts
@@ -1,1 +1,0 @@
-export { default as OccurrenceChip } from './OccurrenceChip';
