+++ src/components/common/index.ts
@@ -1,7 +1,4 @@
-export * from './ConfirmDialog';
+export { default as ConfirmDialog } from './ConfirmDialog';
 export { default as CrossPlatformHorizontalScroll } from './CrossPlatformHorizontalScroll';
-export { default as NoteDialog } from './NoteDialog';
-
-export * from './OccurrenceChip';
-export * from './OccurrenceDialog';
+export { default as ImageCarousel } from './ImageCarousel';
 export { default as VisuallyHiddenInput } from './VisuallyHiddenInput';
