+++ src/components/index.ts
@@ -3,4 +3,6 @@ export * from './calendar';
 export * from './common';
 export * from './habit';
 export * from './header';
+export * from './note';
+export * from './occurrence';
 export * from './trait';
