+++ src/components/note/index.ts
@@ -1,0 +1,1 @@
+export { default as NoteDialog } from './NoteDialog';
