+++ src/components/occurrence/OccurrenceChip.tsx
