+++ src/components/occurrence/index.ts
@@ -1,1 +1,2 @@
+export { default as OccurrenceChip } from './OccurrenceChip';
 export { default as OccurrenceDialog } from './OccurrenceDialog';
